tests: Split up run-strip-reloc.sh test in three subtests
This test can take a very long time under valgrind (on s390x).
Split it into three tests for kernel modules, self-test and ppc64.
That way the tests can run in parallel and each individual
test produces a result (so the buildbot won't time out).

Also change the comparison to elfcmp instead of doing
readelf -a on both and compare the textual output.

	* tests/run-strip-reloc.sh: Split out generic part into...
	* tests/strip-reloc-subr.sh: ...this file.
	* tests/run-strip-reloc-ko.sh: New file.
	* tests/run-strip-reloc-ppc64.sh: New file.
	* tests/run-strip-reloc-self.sh: New file.
	* tests/Makefile.am (TESTS): Remove run-strip-reloc.sh.
	Add run-strip-reloc-ko.sh, run-strip-reloc-self.sh and
	run-strip-reloc-ppc64.sh
	(EXTRA_DIST): Remove run-strip-reloc.sh. Add
	strip-reloc-subr.sh, run-strip-reloc-ko.sh,
	run-strip-reloc-self.sh and run-strip-reloc-ppc64.sh.

Signed-off-by: Mark Wielaard <mark@klomp.org>