PR30962: debuginfod: full paths for X-DEBUGINFOD-FILE/ARCHIVE response headers
Previous code was inconsistent in offering basename versus full
pathname for these headers.  The documentation was not explicit on
this issue.  We now simplify by always passing full names back, and
document this in the debuginfod.8 man page, along with pointers to
how to use proxy front-end servers to strip them if needed.

Signed-Off-By: Frank Ch. Eigler <fche@redhat.com>