PR31248: debuginfod_find_*: lseek to the front on returned fds
Previous code could return a fd that had its file-offset left at the
place where libcurl last used it, namely at the end of the freshly
downloaded file.  Not good if a client just wants to read it right
away!  We now ensure (and document) that the fd is pointed to the
beginning of the file.

Signed-off-by: Frank Ch. Eigler <fche@redhat.com>