libebl: ebl_object_note print 32bit annobin address ranges correctly
Annobin address ranges were always printed as if they were 64bit wide
because addr_size was set to twice the size. This was done because the
note description size should contain two addresses. Fix this by setting
the address size to just one address and then check that descsz is
twice that.

	* libebl/eblobjnote.c (ebl_object_note): Set addr_size to one
	ELF_T_ADDR. Check descsz equals two times addr_size.

Signed-off-by: Mark Wielaard <mark@klomp.org>