libdw: Let clean targets be unconditional
The automake rule "maintainer-clean-generic"
is always available and never conditional,
so let the variable that uses it be define
non-conditionally.

If one actually wants conditional cleaning
they should write a custom rule and set it
as a dependency of a "*clean-local" automake rule.

There is no need to do conditional cleaning here,
so move the MAINTAINERCLEANFILES variable definition
to the end of the Makefile.am file as it is
in the rest of the project.

      * libdw/Makefile.am: move MAINTAINERCLEANFILES
      variable to the end of the file
      as a non-conditional definition.

Signed-off-by: Michael Pratt <mcpratt@pm.me>