configure: better error message for [lib]debuginfod missing dependencies
When dependencies for libdebuginfod, debuginfod or ima verification are
missing and these features are explicitly enabled the user might not
immediately know which of the dependicies are missing. Move the checks
around a little so checks for dependencies are done immediately before
the enable error message. And add the possible reason to the error to
make things more clear.

     * configure.ac: Move libcurl and json-c tests before libdebuginfod
     check, move libmicrohttpd, sqlite3 and libarchive tests before
     debuginfod check and move librpm, libcrypto and imaevm.h tests
     before ima verification check.

https://sourceware.org/bugzilla/show_bug.cgi?id=32294

Signed-off-by: Mark Wielaard <mark@klomp.org>