libdwfl/argp-std.c: Set offline_next_address for a.out default
If none of -e, -p or -k is given as a command line option, libdwfl
parse_opt will use '-e a.out' as the default.

When handling -e, parse_opt will set dwfl->offline_next_address
to 0.  However when handling the default '-e a.out',
dwfl->offline_next_address is not set to 0.  This can result in
eu-addr2line failing to correctly output the line information for a
given address.

Fix this by setting dwfl->offline_next_address to 0 when setting up
the default '-e a.out' dwfl.

https://sourceware.org/bugzilla/show_bug.cgi?id=32538

Signed-off-by: Aaron Merey <amerey@redhat.com>